requirements.txt
Flask>=2.0
Werkzeug>=2.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
